Python Demos/requirements.txt
agent-framework
python-dotenv
azure.identity